+++ docs/create.py
@@ -163,6 +163,9 @@ Notably, all archive files have defined semantic types, which allows us to ensur
 - ``SampleTable[Genotypes]``
     * TSV file for storing target gene's genotype call for each sample.
     * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``.
+- ``SampleTable[Phenotypes]``
+    * TSV file for storing target gene's phenotype call for each sample.
+    * Requires following metadata: ``Gene``, ``SemanticType``.
 - ``SampleTable[Results]``
     * TSV file for storing various results for each sample.
     * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``.